move themes into a feature and cookie helpers to lib
